% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crformats-package.R
\docType{package}
\name{crformats-package}
\alias{crformats}
\alias{crformats-package}
\title{crformats: Using Quarto Extensions Designed for Cancer Registry Reports}
\description{
Provides Quarto extension templates, styles, and helper functions for generating high-quality cancer registry reports.
}
\author{
\strong{Maintainer}: Qiong Chen \email{chenq08@126.com} (\href{https://orcid.org/0000-0003-2401-0046}{ORCID}) (Department of Cancer Epidemiology, The Affiliated Cancer Hospital of Zhengzhou University & Henan Cancer Hospital, Henan Cancer Registry in China) [copyright holder]

}
\keyword{internal}
