% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_docx_template.R
\name{ls_docx}
\alias{ls_docx}
\title{List available docx templates in the package}
\usage{
ls_docx()
}
\value{
A character vector of docx template file names.
}
\description{
List available docx templates in the package
}
\examples{
ls_docx()

}
