% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_ext.R
\name{ls_ext}
\alias{ls_ext}
\title{List bundled Quarto extensions}
\usage{
ls_ext()
}
\value{
a string of available extensions for install
}
\description{
List bundled Quarto extensions
}
\examples{
ls_ext()

}
