% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_docx_template.R
\name{use_docx_template}
\alias{use_docx_template}
\title{Use the docx templates in the manuscript package}
\usage{
use_docx_template(name = "chinese", overwrite = TRUE)
}
\arguments{
\item{name}{Name of the docx template.}

\item{overwrite}{Whether to overwrite the template file if it exists.}
}
\value{
Path of the created/used template.
}
\description{
Use the docx templates in the manuscript package
}
\examples{
\dontrun{
use_docx_template("chinese")
}
}
