% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_ext.R
\name{use_quarto_ext}
\alias{use_quarto_ext}
\title{Install bundled Quarto extensions into specified directory}
\usage{
use_quarto_ext(
  ext_name = "qcreport",
  dir = NULL,
  use_qmd = TRUE,
  qmd_name = "index"
)
}
\arguments{
\item{ext_name}{String indicating which extension to install}

\item{dir}{Character string indicating where extension to install}

\item{use_qmd}{Logical, whether to copy default qmd file}

\item{qmd_name}{Name for the .qmd file (default is 'index.qmd')}
}
\value{
a message if extension was successfully copied over
}
\description{
Install bundled Quarto extensions into specified directory
}
\examples{
\dontrun{
use_quarto_ext("qcreport")
}
}
