% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_docx_template.R
\name{use_docx_template}
\alias{use_docx_template}
\title{Use the docx templates in the manuscript package}
\usage{
use_docx_template(name = "english", overwrite = TRUE)
}
\arguments{
\item{name}{Name of the docx template.}

\item{overwrite}{Whether overwrite the template file if it existed.}
}
\value{
Created docx template.
}
\description{
Use the docx templates in the manuscript package
}
